Fixed bug update coverage add CHANGELOG
