+++ demisto_sdk/commands/format/update_classifier.py
@@ -45,7 +45,8 @@ class OldClassifierJSONFormat(BaseClassifierJSONFormat):
     def run_format(self) -> int:
         try:
             click.secho(f'\n================= Updating file {self.source_file} =================', fg='bright_blue')
-            super().update_json(default_from_version=VERSION_6_0_0)
+            super().update_json()
+            self.set_toVersion()
             self.save_json_to_destination_file()
             return SUCCESS_RETURN_CODE
 
@@ -66,7 +67,7 @@ class ClassifierJSONFormat(BaseClassifierJSONFormat):
     def run_format(self) -> int:
         try:
             click.secho(f'\n================= Updating file {self.source_file} =================', fg='bright_blue')
-            super().update_json()
+            super().update_json(default_from_version=VERSION_6_0_0)
             self.set_description()
             self.set_keyTypeMap()
             self.set_transformer()
